refactor dpm assignment task action routes
