+++ src/app/router_registry.py
@@ -126,6 +126,9 @@ from app.routers.dpm_wave_ai import router as dpm_wave_ai_router
 from app.routers.dpm_wave_campaign_approvals import (
     router as dpm_wave_campaign_approvals_router,
 )
+from app.routers.dpm_wave_campaign_assignment_task_actions import (
+    router as dpm_wave_campaign_assignment_task_actions_router,
+)
 from app.routers.dpm_wave_campaign_assignment_tasks import (
     router as dpm_wave_campaign_assignment_tasks_router,
 )
@@ -333,6 +336,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_campaign_approvals_router)
     app.include_router(dpm_wave_campaign_assignments_router)
     app.include_router(dpm_wave_campaign_assignment_tasks_router)
+    app.include_router(dpm_wave_campaign_assignment_task_actions_router)
     app.include_router(dpm_wave_campaign_workflow_router)
     app.include_router(dpm_construction_router)
     app.include_router(dpm_construction_actions_router)
